fix(encoder): rename tfkeras to keras
