+++ jina/executors/encoders/image/keras.py
